requirements.txt
requests
bs4
sqlalchemy
python-telegram-bot
gevent